python/ai-evaluator-py/requirements.frozen.txt
-e file:///C:/Projects/semantic-kernel-sampler/python/ai-evaluator-py
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
azure-ai-agents==1.1.0
azure-ai-evaluation==1.11.0
azure-ai-projects==1.0.0
azure-core==1.35.0
azure-identity==1.24.0
azure-storage-blob==12.26.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
distro==1.9.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
lagom==2.7.7
markupsafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.6.4
nltk==3.9.1
nodeenv==1.9.1
numpy==2.3.2
oauthlib==3.3.1
openai==1.106.1
pandas==2.3.2
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pyjwt==2.10.1
pyright==1.1.405
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
ruamel-yaml==0.18.15
ruamel-yaml-clib==0.2.12
ruff==0.12.12
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
